feat(lib): Introduces PrayAndPayTestCase class
This commit introduces the `PrayAndPayTestCase` class within the `lib` directory. This base class is designed to group and centralize factory definitions that are intended for reuse across various tests related to the PrayAndPay feature.
